[fix]エネミーの生成データのスクリプトを修正
