+++ Assets/Scripts/Enemy/BaseScripts/EnemyData.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f678d5661c8eafe4ea9eeef937435db5
+guid: 5c8ea5933e8ccb240bdc6392262183c3
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
